ATSPI type validattion
This commit introduces functions for atspi type validation against the XML
descriptions.

This adds and builds on:

Macros

    `test_atspi_event_signature_and_type_match`

Expands to a test function that checks if the signature of an AT-SPI2 event,
comprised of it’s aggregated argument types, match the signature of the
corresponding type in the Rust code, `EventBodyOwned`.

Adds functions:

- `get_signature_of_atspi_event_from_xml`

    Constructs the signature of an AT-SPI2 event from the signal’s arguments in XML.

-  `get_signature_of_method_return_type_from_xml`

    Gets the signature of a method’s return type from XML.

-  ``get_signature_of_signal_body_type`

    Gets the signature of a signal’s return type from XML.